api/requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
google-generativeai==0.1.0